Perform cell-type analysis using partitioned heritability

A common usage of partitioned heritability is to estimate enrichment and significance for an annotation typical for a specific cell-type. For this analysis, it is useful to adjust for a baseline set of annotations, and then estimate the association for a large set of annotations ("cell-types")

Usage

celltype_analysis(sumstat, covariate_dir, ldscore_dir, weights = NULL)

Arguments

sumstat

A dplyr::tibble() with columns SNP, Z and N

covariate_dir

a directory containing the files ld.parquet and annot.parquet

ldscore_dir

filepath to a directory with the annot.parquet and ldscores.parquet files

weights

Optional, a data.frame or tbl with columns SNP, L2

Value

a dplyr::tibble()

Examples

sumstat1 <- arrow::read_parquet(system.file("extdata", "sumstats.parquet", package = "ldsR"))
sumstat1 <- dplyr::select(sumstat1, SNP, Z = Z.x, N=N.x)

celltype_analysis(
 sumstat1,
 covariate_dir = system.file("extdata", "baseline1.1_test", package = "ldsR"),
 ldscore_dir = system.file("extdata", "superclusters", package = "ldsR")
)
#> ℹ Removed 0 rows after merging with ldscores
#> # A tibble: 4 × 8
#>   annot                     coef      coef_se enrich   prop      z   tot tot_se
#>   <chr>                    <dbl>        <dbl>  <dbl>  <dbl>  <dbl> <dbl>  <dbl>
#> 1 amygdala_excitatory    3.02e-8 0.0000000174  0.724 0.112   1.74  0.339 0.0361
#> 2 astrocyte             -1.45e-8 0.0000000160 -0.338 0.0900 -0.911 0.341 0.0387
#> 3 bergmann_glia          4.39e-9 0.0000000154  0.101 0.0881  0.286 0.346 0.0383
#> 4 cerebellar_inhibitory -1.78e-8 0.0000000139 -0.410 0.0874 -1.29  0.343 0.0391