Munge GWAS summary statistics

munge() applies 5 filters (if possible):

Deduplication of RSID

Filter variants on INFO in INFO column is present

Filter variant on effect allele frequency

Removes strand ambigious variants

Removes variants with N < round(stats::quantile(step1$N, 0.9) / 1.5)

Usage

munge(dset, info_filter = 0.9, maf_filter = 0.01)

Arguments

dset

a dplyr::tibble() with columns SNP, A1 A2 Z N and optional columns EAF and INFO

info_filter

INFO score filter threshold at which to remove rows

maf_filter

Minor allele frequenc filter at which to remove rows

Value

a data.frame

Examples

if (FALSE) { # \dontrun{
parse_gwas(tbl)
} # }