Estimate partitioned SNP heritability

An R implementation of the LD score regression method to estimate SNP heritability, mimicking ldsc --h2 from the ldsc package.

For partitioned heritablity, the noMHC weights are used instead of the eur_w_ld_chr weights.

You can inspect the LDscores used by default, in the column L2_celltype arrow::read_parquet(system.file("extdata", "eur_w_ld.parquet", package = "ldsR"))

partitioned_heritability does not perform any quality control on the input summary statistics, except to merge with the SNPs in the reference panel. See the munge() function to mimic the munge_sumstats.py function.

Usage

partitioned_h2(sumstat, ldscore_dir, weights = NULL, n_blocks = 200)

Arguments

sumstat

A dplyr::tibble() with columns SNP, Z and N

ldscore_dir

filepath to a directory with the annot.parquet and ldscores.parquet files

weights

Optional, a data.frame or tbl with columns SNP, L2

n_blocks

Number of blocks to use for the jackknife estimator

Value

a dplyr::tibble()

Examples

sumstat1 <- arrow::read_parquet(system.file("extdata", "sumstats.parquet", package = "ldsR")) |>
dplyr::select(SNP, Z = Z.x, N=N.x)
ldscore_dir <-  system.file("extdata", "baseline1.1_test", package = "ldsR")
fs::dir_tree(ldscore_dir)
#> /home/runner/work/_temp/Library/ldsR/extdata/baseline1.1_test
#> ├── annot.parquet
#> └── ld.parquet
partitioned_h2(sumstat1, ldscore_dir = ldscore_dir)
#> ℹ Removed 449 rows after merging with weights
#> ℹ Removed 5961 rows after merging with ldscores
#> # A tibble: 4 × 8
#>   annot                            coef coef_se enrich   prop     z   tot tot_se
#>   <chr>                           <dbl>   <dbl>  <dbl>  <dbl> <dbl> <dbl>  <dbl>
#> 1 Conserved_LindbladToh.bedL2   8.54e-7 2.55e-7 18.2   0.0212  3.35 0.339 0.0380
#> 2 baseL2                        2.65e-8 8.51e-9  0.563 0.827   3.12 0.339 0.0380
#> 3 H3K4me1_peaks_Trynka.bedL2    6.92e-8 4.66e-8  1.47  0.140   1.48 0.339 0.0380
#> 4 Coding_UCSC.bedL2            -2.28e-7 1.61e-7 -4.85  0.0118 -1.42 0.339 0.0380