Estimate partitioned SNP heritability

An R implementation of the LD score regression method to estimate SNP heritability, mimicking ldsc --h2 from the ldsc package.

For partitioned heritablity, the noMHC weights are used instead of the eur_w_ld_chr weights.

You can inspect the LDscores used by default, in the column L2_celltype arrow::read_parquet(system.file("extdata", "eur_w_ld.parquet", package = "ldsR"))

partitioned_heritability does not perform any quality control on the input summary statistics, except to merge with the SNPs in the reference panel. See the munge() function to mimic the munge_sumstats.py function.

Usage

partitioned_heritability(sumstat, ldscore_dir, weights = NULL, n_blocks = 200)

Arguments

sumstat

A dplyr::tibble() with columns SNP, Z and N

ldscore_dir

filepath to a directory with the annot.parquet and ldscores.parquet files

weights

Optional, a data.frame or tbl with columns SNP, L2

n_blocks

Number of blocks to use for the jackknife estimator

Value

a dplyr::tibble()

Examples

if (FALSE) { # \dontrun{
partitioned_heritability(sumstats, "path/to_dir/")
} # }